Setup base HTML and JavaScript file.
Setup the index.html file to include a form to represent the search bar and a section where the results will by dynamically rendered. Add event listener to the form, where upon submission, it will hit the target Twitch API route. Throw an alert to notify the user that the route was hit and console.log out the XHR response text.