Add pagination 'menu', implement next page of results functionality.
Added a navbar for the search results section of the page. The navbar contains the total number of results found in addition to the newly added controls for going back and forth between the results pages if applicable. Only implemented the next results page functionality; it redirects to the actual API Route, will work on it to dynamically refresh the results section of the page before implementing the previous results page functionality.