Dynamically change current page index upon hitting next and previous buttons.
Used the XHR response object to retrieve the limit and offset parameter values which correspond to the number of results show per page and 'current set' of results. Modified the HTML to have two spans within a paragraph that shows the page index, where the first reveals the current page index, and the latter refers to the total number of pages. Also identified a potential bug with the API: the response still contains a 'next link' for the 'last set ' of results, which will redirect you to that route even if it contains 0 streams (the streams array/collection would be empty).