Address next button redirect issue.
The next button would hit one of the API routes even when a user was looking at the last page of results from their search. As mentioned in an earlier commit, the XHR response still provides a 'next link', but upon redirecting to that next link, there would be no stream objects to render. The current fix checks to see if the user is on the last page, and if so, prevents the user from being redirected. Next step is to toggle the visibility of the previous and next buttons based on more conditional logic.