Toggle display of pagination buttons.
Added a displ-none class to the pagination buttons and the results message elements so that when the user first loads the page, the user does not see them since no search has been made yet. Based on whether or not the route hit contains a previous and/or next link, their respective buttons will be toggled to be visible. Considered simple using JavaScript's classList.toggle() approach, but that seemed less efficient when I had to consider what the classList state of the buttons were after each page redirection. Chose to go with manipulating the style rather than toggling classes and writing more lines of code.