Refactor code to follow DRY principles.
Changed the navigateToPage function to take in a second parameter, url, which would point to the target API route. Implemented a ternary in order to set a default value for the url parameter. The use case for the ternary applies when a user clicks either the previous or next button, where rather than passing in a url, we default the undefined url variable to be equal to the button's parent's href attribute. Could have alternatively set the value to be equal to the event target's href property possibly, but I wanted to practice using ternaries.