Refactor and reorganize JavaScript code.
Mainly reorganized code to have functions listed in the 'order that they are executed' for readability.
Identified issue in logic: when a user receives a set of results greater than 100, the page index resets to numbers 1-9 rather than moving onto 10 because the current logic only takes the first value of the offset parameter.