Add loading animation between searches/pagination.
Included loading image to be rendered in between requests with a CSS animation that causes the image to shift left and right. The image is appended to the results section of the page when a search request is made or when a pagination button is clicked on. Once the request is complete, the image is removed.