Finish initial implementation of tooltips.
Upon stream description hover event, page renders a tooltip containing the respective channel name and channel logo. The tooltip is clickable, and upon clicking one, the user's window will open a new tab that links to the actual channel on Twitch. The page also becomes tinted when a tooltip is active, and returns to normal once the user moves their mouse away from the tooltip and stream description. Something to think about is whether or not the tooltip is intuitive, as it may worsen a user's experience given that the user may have to go out of their way to move their mouse to continue navigating the page.